Refactor and re-enable test(s)
